% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_spreadsheets.R
\name{check_folder}
\alias{check_folder}
\title{Check Folder}
\usage{
check_folder(folder = NULL)
}
\arguments{
\item{folder}{a string with a folder name. Defaults to \code{NULL}}
}
\description{
Check whether a folder exists and if so whether it should be
  overwritten.
}
