% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_RDCOMClient}
\alias{install_RDCOMClient}
\title{Install RDCOMClient}
\usage{
install_RDCOMClient()
}
\value{
thing
}
\description{
RDCOMClient is a necessary package for the package to interact with 
  Excel spreadsheets.
}
\details{
Installs RDCOMClient from Github
}
