% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_results}
\alias{read_results}
\title{Read results}
\usage{
read_results(folder_directory = getwd())
}
\arguments{
\item{folder_directory}{directory to search for folder. Defaults to \code{getwd()}}
}
\description{
Read the results files that are in .csv format
}
