% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_interact_excel_2.R
\name{update_spreadsheet_2}
\alias{update_spreadsheet_2}
\title{R_interact_excel 2}
\usage{
update_spreadsheet_2(file_location = NULL, sheet = "Sheet1",
  input_values = NULL, input_cells = NULL, output_cells = NULL,
  save_spreadsheet = "new_spreadsheet.xls", wb = NULL)
}
\arguments{
\item{file_location}{file location of the spreadsheet to modify. The
spreadsheet should be \code{.xls} format}

\item{sheet}{the sheet name. Defaults to \code{Sheet1}}

\item{input_values}{a vector of values corresponding to the cells to be modified}

\item{input_cells}{a character vector of the input cells 
in the form c('A1', 'A2')}

\item{output_cells}{the cells from which the output will be read}

\item{save_spreadsheet}{if ' ' then will overwrite the current spreadsheet.
If \code{NULL} will not save. If any other name, will save the file as an
\code{.xls} object}

\item{wb}{a spreadsheet}
}
\description{
Allows a spreadsheet to be modified programatically and the 
  outputs captured. Uses the \code{RDCOMClient}, \code{dplyr} and
  \code{stringr} packages
}
\examples{
\dontrun{
input_cells <- c('A1', 'A2')
input_values = c(4, 9)
output_cells = c('B1', 'B2')
# ADD SAMPLE SPREADSHEET TO PACKAGE AND RUN CODE
}
}
