% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_spreadsheets_2.R
\name{use_run_matrix_2}
\alias{use_run_matrix_2}
\title{Multiple Spreadsheets}
\usage{
use_run_matrix_2(run_matrix = NULL, file_location = NULL,
  sheet = "Sheet1", folder = "results", round_dp = 3)
}
\arguments{
\item{run_matrix}{a file path to the input matrix OR the matrix itself}

\item{file_location}{the full path of the calculation spreadsheet}

\item{sheet}{the sheet in the spreadsheet in which you wish to change values}

\item{folder}{the folder to save the results to}

\item{round_dp}{the number of significant figures to round to. Defaults to 3}
}
\description{
Make changes to a spreadsheet and extract the outputs from 
  specified cells using a run matrix as an input.
}
